examples/planets/src/requirements.txt
black==21.5b1
flake8>=3.7.9, <4.0
ipython==7.10
isort~=5.0
jupyter~=1.0
jupyter_client>=5.1, <7.0
jupyterlab~=3.0
kedro==0.17.4
nbstripout~=0.4
pytest-cov~=2.5
pytest-mock>=1.7.1, <2.0
pytest~=6.2
wheel>=0.35, <0.37
kedro-neptune
neptune-client[optuna,fastai]>=0.10.2
